requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-crontab==3.0.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1 